fs.promises

+++ FS/async/10-rmdir/index.js
@@ -2,22 +2,26 @@ const fs = require("fs");
 
 /**
  * 读取目录
- * @params path 目录路径
- * @params callback 回调函数 => error-first
+ * @param {string} 目录路径
+ * @param {function} 回调函数 => error-first
  */
 fs.readdir("./test", (err, res) => {
   if (err) {
     console.log(err);
   } else {
-    res.forEach((item) => {
-      // 异步方法，可能会出现文件未全部删除就删除文件夹的问题
-      fs.unlink(`./test/${item}`, (err) => {
-        if (err) {
-          console.log(err);
-        }
+    // 删除文件
+    if (res.length > 0) {
+      res.forEach((item) => {
+        // 异步方法，存在文件未全部删除就删除文件夹，从而发生报错的问题
+        fs.unlink(`./test/${item}`, (err) => {
+          if (err) {
+            console.log(err);
+          }
+        });
       });
-    });
+    }
 
+    // 删除目录
     fs.rmdir("./test", (err) => {
       if (err) {
         console.log(err);
